requirements.txt
flask
tensorflow
opencv-python
mediapipe
numpy
pillow